simplify file extensions 'js.jsx' -> '.jsx'
