+++ web-experimental/src/App.jsx
@@ -1,15 +1,18 @@
 import React from 'react';
 import { Redirect } from 'react-router';
+import { Route } from 'react-router-dom';
 import './App.css';
-import ProductRouter from './resources/product/ProductRouter.js.jsx';
-import ProductRouter2 from './resources/product2/ProductRouter.js.jsx';
-import UserRouter from './resources/user/UserRouter.js.jsx';
+import ProductRouter from './resources/product/ProductRouter.jsx';
+import ProductRouter2 from './resources/product2/ProductRouter.jsx';
+import UserRouter from './resources/user/UserRouter.jsx';
 function App() {
   return (
     <>
       {/* import all resource routers here */}
       {/* redirect for now since there is no view for the base route */}
-      <Redirect to="/products2"/>
+      <Route exact path='/'>
+        <Redirect to="/products2" />
+      </Route>      
       <ProductRouter />
       <ProductRouter2 />
       <UserRouter />
